“laravel_pj”

+++ database/migrations/2022_04_10_080338_create_matchings_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
     {
         Schema::create('matchings', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('post_id');
-            $table->integer('request_user_id');
+            $table->unsignedBigInteger('post_id');
+            $table->unsignedBigInteger('request_user_id');
             $table->boolean('is_match');
         });
     }
